Allow OneGenOf to work in struct fields
This is similar to #56.  OneGenOf doesn't work well in struct fields
because the Struct() Gen uses a GenResult to determine the type of the
field, rather than using the field's declared type.  This means that,
when OneGenOf returns subsequent values whose type doesn't match, it
causes a runtime type error and a panic.

The fix is to use the field's declared type as the basis for the
Upstream and Downstream functions in BiMapper.

To take a concrete example:

    type Struct struct {
        Field interface{}
    }
    gen.Struct(reflect.TypeOf(&Struct{}), map[string]gopter.Gen{
        "Field": gen.OneGenOf(gen.AnyString(), gen.Int())
    }

In this generator, Struct will pull a sample from OneGenOf and get
either a string or an int.  For the sake of argument, suppose it gets a
string: then it assumes that all subsequent values will be a string.  It
then builds a BiMapper using functions of down type `func(string) Struct`
and up type `func(Struct) string)`.  This breaks as soon as the OneGenOf
generator emits an int instead.

The fix in this commit is to use the field's declared type in the
function types in the BiMapper.  In the above example, it means
constructing functions of type `func(interface{}) Struct` and
`func(Struct) interface{}`.

It turns out that #55 misdiagnosed the problem it was trying to solve,
so I've had to revert the code changes there (though I've kept the
tests, which still pass).